make small ACPI implemantation work in both loader and kernel

+++ src/kernel/arch/x86_64/interrupts/APICEnum.hpp
@@ -4,16 +4,16 @@
 
 namespace apic {
 
-enum class DeliveryMode : uint32_t {
+enum class DeliveryMode : uint8_t {
     FIXED, LOWEST_PRIORITY, SMI, RESERVED1, NMI, INIT, STARTUP, RESERVED2
 };
-enum class TriggerMode : uint32_t {
+enum class TriggerMode : uint8_t {
     EDGE, LEVEL
 };
-enum class Polarity : uint32_t {
+enum class Polarity : uint8_t {
     ACRIVE_HIGH, ACTIVE_LOW
 };
-enum class Level : uint32_t {
+enum class Level : uint8_t {
     DEASSERT, ASSERT
 };
